refactor: move PacketIds from static class to proto enum
PacketIds is now defined as a proto enum in template-projects/Protos/PacketIds.proto,
co-located with Sample.proto. Generated C# enum lives in the same
FastPortGameServerTemplate.Protocols namespace as EchoRequest/EchoResponse.

The C# protobuf generator strips the PACKET_IDS_ enum-name prefix automatically,
so call sites still see `PacketIds.EchoRequest` / `PacketIds.EchoResponse`
(now of type PacketIds enum, not int constants).

Removed:
- template-projects/FastPortGameServerTemplate/Handlers/PacketIds.cs
- SampleClient.csproj's <Compile Include Link> for PacketIds.cs
  (no longer needed; SampleClient gets the enum from its own <Protobuf Include>)

Added:
- template-projects/Protos/PacketIds.proto with PACKET_IDS_ECHO_REQUEST=1001,
  PACKET_IDS_ECHO_RESPONSE=1002 plus required _UNSPECIFIED=0 zero-value.
- (int) cast at 5 call sites in EchoHandler.cs (server) and
  SampleClientSession.cs (client) — enum-to-int is not implicit in C#.

sln updates:
- FastPortSharp.sln: register PacketIds.proto under Protos Solution Folder
- scaffold output sln: inject_protos_solution_folder() already enumerates
  Protos/*.proto, so PacketIds.proto appears automatically.

Fixtures:
- case-01/05/07 files-present.txt: remove .../Handlers/PacketIds.cs entry,
  add Protos/PacketIds.proto entry.
- case-01 sha256/tree regenerated via --update-golden.

Docs:
- Template README + QUICKSTART: explain enum + (int) cast pattern.
- Top-level README.md + README.ko.md: tree diagrams reflect PacketIds.proto.

Verified:
- dotnet build FastPortSharp.sln: 0 error / 0 warning
- dotnet test: 139 (Engine) + 37 (Dashboard) = 176 PASS
- Echo round-trip: RTT=15.670ms
- run.sh sh + ps1: 7/7 PASS each

+++ template-projects/FastPortGameServerTemplate.SampleClient/Sessions/SampleClientSession.cs
@@ -1,4 +1,3 @@
-using FastPortGameServerTemplate.Handlers;
 using FastPortGameServerTemplate.Protocols;
 using LibCommons;
 using LibNetworks.Extensions;
@@ -39,7 +38,7 @@ public sealed class SampleClientSession : BaseSessionServer
 
         m_SendTimestamp = Stopwatch.GetTimestamp();
         var request = new EchoRequest { Message = m_Options.Message };
-        RequestSendMessage(PacketIds.EchoRequest, request);
+        RequestSendMessage((int)PacketIds.EchoRequest, request);
     }
 
     protected override void OnReceived(BasePacket packet)
@@ -52,9 +51,9 @@ public sealed class SampleClientSession : BaseSessionServer
             return;
         }
 
-        if (packetId != PacketIds.EchoResponse)
+        if (packetId != (int)PacketIds.EchoResponse)
         {
-            m_Logger.LogWarning("Unexpected packet id. Expected={Expected}, Got={Actual}", PacketIds.EchoResponse, packetId);
+            m_Logger.LogWarning("Unexpected packet id. Expected={Expected}, Got={Actual}", (int)PacketIds.EchoResponse, packetId);
             return;
         }
